requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
python-dotenv
email_validator
fpdf
